old_counseling_linebot/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cheroot==10.0.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.functools==4.1.0
Jinja2==3.1.5
jiter==0.8.2
line-bot-sdk==3.14.5
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.1.0
openai==1.60.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.11.1
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
